+++ src/widgets/input.rs
@@ -1199,6 +1199,10 @@ impl Widget for Input {
     fn styles_mut(&mut self) -> Option<&mut WidgetStyles> {
         Some(&mut self.styles)
     }
+
+    fn get_selection(&self) -> Option<String> {
+        self.selected_text()
+    }
 }
 
 impl Renderable for Input {